pd: Always set number of DSP inlets to >= 1
Pure data implicitly registers the first inlet, even when it is not
used. That causes an unexpected offset of output buffers, where we read
the last input buffer instead of the first output.

With this change, we enforce that the number is always set to >= 1 to
account for this inlet.

Signed-off-by: Petr Horáček <phoracek@redhat.com>